Move stringlist_select_matching_files to rd_util.cpp
